refactor: dispatch styling component from toolbar
