+++ src/App.tsx
@@ -8,6 +8,7 @@ import EditFeatureDrawer from './components/EditFeatureDrawer';
 import Footer from './components/Footer';
 import Header from './components/Header';
 import LayerDetailsModal from './components/LayerDetailsModal';
+import StylingDrawer from './components/StylingDrawer';
 import ToolMenu from './components/ToolMenu';
 import UploadDataModal from './components/UploadDataModal';
 
@@ -44,8 +45,7 @@ export const App: React.FC<AppProps> = ({
       <UploadDataModal />
       <EditFeatureDrawer />
       <LayerDetailsModal />
-      {/* // todo: move styling component here and use store value for visibility
-      //todo: properties an componmente: open */}
+      <StylingDrawer />
     </div>
   );
 };
