+++ src/components/ToolMenu/Draw/StylingDrawerButton/index.spec.tsx
